+++ Snowflake.Data.Tests/UnitTests/Revocation/CrlTest.cs
@@ -1,17 +1,21 @@
 using System;
 using System.IO;
-using System.Security.Cryptography;
-using System.Security.Cryptography.X509Certificates;
 using Moq;
-using Org.BouncyCastle.Asn1.X509;
-using Org.BouncyCastle.Crypto.Operators;
 using Xunit;
 using Org.BouncyCastle.Math;
-using Org.BouncyCastle.Security;
-using Org.BouncyCastle.X509;
+using Snowflake.Data.Configuration;
 using Snowflake.Data.Core.Revocation;
 using Snowflake.Data.Core.Tools;
 using Snowflake.Data.Tests.Util;
+
+#if NET8_0_OR_GREATER
+using System.Security.Cryptography;
+using System.Security.Cryptography.X509Certificates;
+using Org.BouncyCastle.Asn1.X509;
+using Org.BouncyCastle.Crypto.Operators;
+using Org.BouncyCastle.Security;
+using Org.BouncyCastle.X509;
+#endif
 
 namespace Snowflake.Data.Tests.UnitTests.Revocation
 {
@@ -22,8 +26,8 @@ namespace Snowflake.Data.Tests.UnitTests.Revocation
         {
             // arrange
             var crlBytes = File.ReadAllBytes(s_digiCertCrlPath);
-            var environmentOperations = new Mock<EnvironmentOperations>();
-            var crlParser = new CrlParser(environmentOperations.Object);
+            var envFacade = new Mock<IEnvironmentFacade>();
+            var crlParser = new CrlParser(envFacade.Object);
             var now = new DateTime(2025, 7, 25, 16, 57, 0, DateTimeKind.Utc);
             var expectedCrlDistributionPoints = new[] { DigiCertCrlUrl1, DigiCertCrlUrl2 };
 
@@ -54,8 +58,8 @@ namespace Snowflake.Data.Tests.UnitTests.Revocation
         {
             // arrange
             var crlBytes = File.ReadAllBytes(s_digiCertCrlPath);
-            var environmentOperations = new Mock<EnvironmentOperations>();
-            var crlParser = new CrlParser(environmentOperations.Object);
+            var envFacade = new Mock<IEnvironmentFacade>();
+            var crlParser = new CrlParser(envFacade.Object);
             var now = DateTimeOffset.Parse(nowString).UtcDateTime;
             var crl = crlParser.Parse(crlBytes, now);
 
@@ -70,8 +74,10 @@ namespace Snowflake.Data.Tests.UnitTests.Revocation
         public void TestCrlParserDefaultValidityTime()
         {
             // arrange
-            var environmentOperations = new Mock<EnvironmentOperations>();
-            var crlParser = new CrlParser(environmentOperations.Object);
+            var envFacade = new Mock<IEnvironmentFacade>();
+            envFacade.Setup(e => e.GetInt(EnvVars.CrlValidityTime)).Returns(EnvVars.CrlValidityTime.DefaultValue);
+            envFacade.Setup(e => e.GetInt(EnvVars.CrlCacheRemovalDelay)).Returns(EnvVars.CrlCacheRemovalDelay.DefaultValue);
+            var crlParser = new CrlParser(envFacade.Object);
 
             // act
             var validityTime = crlParser.GetCacheValidityTime();
@@ -84,11 +90,11 @@ namespace Snowflake.Data.Tests.UnitTests.Revocation
         public void TestCrlParserCustomValidityTime()
         {
             // arrange
-            var environmentOperations = new Mock<EnvironmentOperations>();
-            environmentOperations
-                .Setup(e => e.GetEnvironmentVariable(CrlParser.CrlValidityTimeEnvName))
-                .Returns("7");
-            var crlParser = new CrlParser(environmentOperations.Object);
+            var envFacade = new Mock<IEnvironmentFacade>();
+            envFacade
+                .Setup(e => e.GetInt(EnvVars.CrlValidityTime))
+                .Returns(7);
+            var crlParser = new CrlParser(envFacade.Object);
 
             // act
             var validityTime = crlParser.GetCacheValidityTime();
@@ -102,8 +108,8 @@ namespace Snowflake.Data.Tests.UnitTests.Revocation
         {
             // arrange
             var crlBytes = File.ReadAllBytes(s_digiCertCrlPath);
-            var environmentOperations = new Mock<EnvironmentOperations>();
-            var crlParser = new CrlParser(environmentOperations.Object);
+            var envFacade = new Mock<IEnvironmentFacade>();
+            var crlParser = new CrlParser(envFacade.Object);
             var now = new DateTime(2025, 7, 25, 16, 57, 0, DateTimeKind.Utc);
             var crl = crlParser.Parse(crlBytes, now);
             var cacheValidityTime = TimeSpan.FromDays(1);
@@ -128,8 +134,8 @@ namespace Snowflake.Data.Tests.UnitTests.Revocation
         {
             // arrange
             var bigInt = new BigInteger(stringValue);
-            var environmentOperations = new Mock<EnvironmentOperations>();
-            var crlParser = new CrlParser(environmentOperations.Object);
+            var envFacade = new Mock<IEnvironmentFacade>();
+            var crlParser = new CrlParser(envFacade.Object);
 
             // act
             var hexString = crlParser.ConvertToHexadecimalString(bigInt);
@@ -146,8 +152,8 @@ namespace Snowflake.Data.Tests.UnitTests.Revocation
         {
             // arrange
             var crlBytes = File.ReadAllBytes(s_digiCertCrlPath);
-            var environmentOperations = new Mock<EnvironmentOperations>();
-            var crlParser = new CrlParser(environmentOperations.Object);
+            var envFacade = new Mock<IEnvironmentFacade>();
+            var crlParser = new CrlParser(envFacade.Object);
             var now = new DateTime(2010, 04, 10, 16, 57, 0, DateTimeKind.Utc);
             var crl = crlParser.Parse(crlBytes, now);
 
@@ -201,8 +207,8 @@ namespace Snowflake.Data.Tests.UnitTests.Revocation
             var bouncyCrl = crlGenerator.Generate(signatureFactory);
 
             // Parse through production code path
-            var environmentOperations = new Mock<EnvironmentOperations>();
-            var crlParser = new CrlParser(environmentOperations.Object);
+            var envFacade = new Mock<IEnvironmentFacade>();
+            var crlParser = new CrlParser(envFacade.Object);
             var crl = crlParser.Create(bouncyCrl, now);
 
             var bouncyCastleHex = crlParser.ConvertToHexadecimalString(serialAsBigInt);