+++ Snowflake.Data/Client/SnowflakeDbConnection.cs
@@ -195,16 +195,28 @@ namespace Snowflake.Data.Client
         public override void Close()
         {
             logger.Debug("Close Connection.");
-            if (IsNonClosedWithSession())
+
+            try
             {
-                var returnedToPool = TryToReturnSessionToPool();
-                if (!returnedToPool)
+                if (IsNonClosedWithSession())
                 {
-                    SfSession.close();
+                    var returnedToPool = TryToReturnSessionToPool();
+                    if (!returnedToPool)
+                    {
+                        SfSession.close();
+                    }
+
+                    SfSession = null;
                 }
-                SfSession = null;
+
+                _connectionState = ConnectionState.Closed;
             }
-            _connectionState = ConnectionState.Closed;
+            catch (Exception exception)
+            {
+                logger.Error($"Close Connection failed: {exception.Message}");
+                _connectionState = ConnectionState.Broken;
+                throw;
+            }
         }
 
 #if NETCOREAPP3_0_OR_GREATER
@@ -219,55 +231,43 @@ namespace Snowflake.Data.Client
         public virtual async Task CloseAsync(CancellationToken cancellationToken)
         {
             logger.Debug("Close Connection.");
-            TaskCompletionSource<object> taskCompletionSource = new TaskCompletionSource<object>();
 
             if (cancellationToken.IsCancellationRequested)
+                return;
+
+            if (!IsNonClosedWithSession())
             {
-                taskCompletionSource.SetCanceled();
+                logger.Debug("Session not opened. Nothing to do.");
+                return;
             }
-            else
+
+            var returnedToPool = TryToReturnSessionToPool();
+            if (returnedToPool)
             {
-                if (IsNonClosedWithSession())
-                {
-                    var returnedToPool = TryToReturnSessionToPool();
-                    if (returnedToPool)
-                    {
-                        _connectionState = ConnectionState.Closed;
-                        taskCompletionSource.SetResult(null);
-                    }
-                    else
-                    {
-                        await SfSession.CloseAsync(cancellationToken).ContinueWith(
-                            previousTask =>
-                            {
-                                if (previousTask.IsFaulted)
-                                {
-                                    // Exception from SfSession.CloseAsync
-                                    logger.Error("Error closing the session", previousTask.Exception);
-                                    taskCompletionSource.SetException(previousTask.Exception);
-                                }
-                                else if (previousTask.IsCanceled)
-                                {
-                                    _connectionState = ConnectionState.Closed;
-                                    logger.Debug("Session close canceled");
-                                    taskCompletionSource.SetCanceled();
-                                }
-                                else
-                                {
-                                    logger.Debug("Session closed successfully");
-                                    _connectionState = ConnectionState.Closed;
-                                    taskCompletionSource.SetResult(null);
-                                }
-                            }, cancellationToken).ConfigureAwait(false);
-                    }
-                }
-                else
-                {
-                    logger.Debug("Session not opened. Nothing to do.");
-                    taskCompletionSource.SetResult(null);
-                }
+                _connectionState = ConnectionState.Closed;
+                SfSession = null;
+                return;
             }
-            await taskCompletionSource.Task.ConfigureAwait(false);
+
+            try
+            {
+                await SfSession.CloseAsync(cancellationToken).ConfigureAwait(false);
+                logger.Debug("Session closed successfully");
+                _connectionState = ConnectionState.Closed;
+                SfSession = null;
+            }
+            catch (OperationCanceledException)
+            {
+                _connectionState = ConnectionState.Closed;
+                logger.Debug("Session close canceled");
+                throw;
+            }
+            catch (Exception exception)
+            {
+                _connectionState = ConnectionState.Broken;
+                logger.Error("Error closing the session", exception);
+                throw;
+            }
         }
 
         protected virtual bool CanReuseSession(TransactionRollbackStatus transactionRollbackStatus)
@@ -286,8 +286,8 @@ namespace Snowflake.Data.Client
             }
             try
             {
-                FillConnectionStringFromTomlConfigIfNotSet();
                 OnSessionConnecting();
+                FillConnectionStringFromTomlConfigIfNotSet();
                 var sessionContext = new SessionPropertiesContext
                 {
                     Password = Password,
@@ -324,16 +324,14 @@ namespace Snowflake.Data.Client
                 ConnectionString = _tomlConnectionBuilder.GetConnectionStringFromToml();
         }
 
-        public override Task OpenAsync(CancellationToken cancellationToken)
+        public override async Task OpenAsync(CancellationToken cancellationToken)
         {
             logger.Debug("Open Connection Async.");
             if (_connectionState != ConnectionState.Closed)
             {
                 logger.Debug($"Open with a connection already opened: {_connectionState}");
-                return Task.CompletedTask;
+                return;
             }
-            OnSessionConnecting();
-            FillConnectionStringFromTomlConfigIfNotSet();
             var sessionContext = new SessionPropertiesContext
             {
                 Password = Password,
@@ -342,36 +340,31 @@ namespace Snowflake.Data.Client
                 Token = Token
             };
 
-            return SnowflakeDbConnectionPool
-                .GetSessionAsync(ConnectionString, sessionContext, cancellationToken)
-                .ContinueWith(previousTask =>
-                {
-                    if (previousTask.IsFaulted)
-                    {
-                        // Exception from SfSession.OpenAsync
-                        Exception sfSessionEx = previousTask.Exception;
-                        _connectionState = ConnectionState.Closed;
-                        logger.Error("Unable to connect", sfSessionEx);
-                        throw new SnowflakeDbException(
-                            sfSessionEx,
-                            SnowflakeDbException.CONNECTION_FAILURE_SSTATE,
-                            SFError.INTERNAL_ERROR,
-                            "Unable to connect");
-                    }
-                    else if (previousTask.IsCanceled)
-                    {
-                        _connectionState = ConnectionState.Closed;
-                        logger.Debug("Connection canceled");
-                        throw new TaskCanceledException("Connecting was cancelled");
-                    }
-                    else
-                    {
-                        // Only continue if the session was opened successfully
-                        SfSession = previousTask.Result;
-                        logger.Debug($"Connection open with pooled session: {SfSession.sessionId}");
-                        OnSessionEstablished();
-                    }
-                }, TaskContinuationOptions.None); // this continuation should be executed always (even if the whole operation was canceled) because it sets the proper state of the connection
+            try
+            {
+                OnSessionConnecting();
+                FillConnectionStringFromTomlConfigIfNotSet();
+                var session = await SnowflakeDbConnectionPool.GetSessionAsync(ConnectionString, sessionContext, cancellationToken).ConfigureAwait(false);
+                SfSession = session;
+                logger.Debug($"Connection open with pooled session: {SfSession.sessionId}");
+                OnSessionEstablished();
+            }
+            catch (OperationCanceledException)
+            {
+                _connectionState = ConnectionState.Closed;
+                logger.Debug("Connection canceled");
+                throw new TaskCanceledException("Connecting was cancelled");
+            }
+            catch (Exception ex)
+            {
+                _connectionState = ConnectionState.Closed;
+                logger.Error("Unable to connect", ex);
+
+                if (ex is SnowflakeDbException)
+                    throw;
+
+                throw new SnowflakeDbException(ex, SnowflakeDbException.CONNECTION_FAILURE_SSTATE, SFError.INTERNAL_ERROR, "Unable to connect");
+            }
         }
 
         public Mutex GetArrayBindingMutex()